requirements.txt
python-dotenv
numpy
torch
torchvision
scikit-learn
umap-learn

matplotlib
seaborn
tqdm
neptune

pillow
pandas

ipython>=8.0.0,<9.0.0
ipykernel
